GenAI_SourceCode/19_movie_recommend/teams/requirements.txt
botbuilder-core
asyncio
fastapi
uvicorn
requests
python-dotenv
loguru
aiohttp